made some implicit casts explicit

+++ code/cmd_chop.cc
@@ -140,7 +140,7 @@ static int chopHit(TBeing *c, TBeing *v, int score)
       act("The spikes on $n's $o sink into $N.", FALSE, c, item, v, TO_NOTVICT);
       act("The spikes on $n's $o sink into you.", FALSE, c, item, v, TO_VICT);
 
-      if (c->reconcileDamage(v, dam*0.15, TYPE_STAB) == -1)
+      if (c->reconcileDamage(v, (int)(dam*0.15), TYPE_STAB) == -1)
 	return DELETE_VICT;
     }
